ml/requirements.txt
Pillow>=10.4.0
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.45.0
